+++ fiddles/three/fiddle-0023-Jupiter/src/metadata.js
@@ -1,0 +1,8 @@
+let metadata = {
+  urls: {
+    github: 'https://github.com/bradyhouse/house/tree/master/fiddles/three/fiddle-0023-Jupiter',
+    jupiter: {
+      surfaceMaterial: 'https://s3-us-west-2.amazonaws.com/s.cdpn.io/297733/jupiterSurfaceMaterial.jpg'
+    }
+  }
+};